Move Game into core!
WIth this, Game does not depend on anything from the browser environment, though there's still a bunch of initialization code in the client that should live in core.
